bots/discord/requirements.txt
discord.py>=2.0,<3.0
python-dotenv>=1.0
